% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{Read the FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character vector giving the path to the file containing the
dataset}
}
\value{
This function returns a data frame parsed from the given dataset
}
\description{
This function reads in one of the Fatality Analysis Reporting System datasets
provided by the US National Highway Traffic Safety Administration and returns
it with a \code{\link{tbl_df}} class.
This function will stop with an error if the filename does not exist.
}
\examples{
fars_read('data/accident_2013.csv.bz2')
fars_read(make_filename(2015))

}
