% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load months and years of FARS records for given year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{one or more years which is coercable to an integer with the
\code{\link{as.integer}} function}
}
\value{
a list of data tables, one for each value in years
}
\description{
This function will stop with an error if the a dataset for the year does not
exist in the working directory or if any empty vector is used.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013, 2015))
}
