% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Generate table of FARS records per month for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{one or more years which is coercable to an integer with the
\code{\link{as.integer}} function}
}
\value{
a data.frame containing a column labeled MONTH (with values ranging
  from 1 to 12) and another column for each year listing the totals per month.
}
\description{
This function generates a summary table giving, for each month, the number of
records with the Fatality Analysis Reporting System dataset for that year. This
is arranged so that the results for each year are reported in a single column.
}
\details{
This function will stop with an error if the a dataset for the year does not
exist in the working directory or if any empty vector is used.
}
\examples{
fars_summarize_years(2013) # display summary for 2013
fars_summarize_years(c(2013,2015)) # show 2013 and 2015 data

}
